fix ordering for termini of nested internal-nodes
Expressions can be nested inside each other in the data returned by
getParseData() in such a way that the end position of the nested
expressions are identical.

For example, `equal_assign` and `expr[2]` end at the same position
in the following (relevant in R3.6).
- equal_assign:
    - expr[1]:
        - SOME_SYMBOL
    - EQ_ASSIGN
    - expr[2]:
        - SOME_VALUE

xml-termini for `equal_assign` and both `expr` nodes are added
while running xml_parse_data; but these termini were originally
sorted such that that for `equal_assign` came before the second
`expr`. Like: `<equal_assign>...<expr[2]></equal_assign></expr[2]>`.

The xml-termini for each expression are now sorted by the start
position of the expressions (in addition to their other
coordinates).